LayoutTests/imported/w3c/web-platform-tests/webcodecs/video-encoder-config.https.any.worker-expected.txt


PASS Test that VideoEncoder.isConfigSupported() rejects invalid config:Emtpy codec
PASS Test that VideoEncoder.isConfigSupported() rejects invalid config:Unrecognized codec
FAIL Test that VideoEncoder.isConfigSupported() rejects invalid config:Width is too large assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Test that VideoEncoder.isConfigSupported() rejects invalid config:Height is too large assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Test that VideoEncoder.isConfigSupported() rejects invalid config:Invalid scalability mode assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Test that VideoEncoder.isConfigSupported() rejects invalid config:AVC not supported by VP8 assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL VideoEncoder.isConfigSupported() doesn't support config:Too strenuous accelerated encoding parameters assert_false: expected false got true
FAIL VideoEncoder.isConfigSupported() doesn't support config:Odd sized frames for H264 assert_false: expected false got true
PASS VideoEncoder.isConfigSupported() supports:{"codec":"avc1.42001E","hardwareAcceleration":"no-preference","width":640,"height":480,"bitrate":5000000,"framerate":24,"avc":{"format":"annexb"},"futureConfigFeature":"foo"}
PASS VideoEncoder.isConfigSupported() supports:{"codec":"vp8","hardwareAcceleration":"no-preference","width":800,"height":600,"bitrate":7000000,"bitrateMode":"variable","framerate":60,"scalabilityMode":"L1T2","futureConfigFeature":"foo","latencyMode":"quality"}
FAIL VideoEncoder.isConfigSupported() supports:{"codec":"vp09.00.10.08","hardwareAcceleration":"no-preference","width":1280,"height":720,"bitrate":7000000,"bitrateMode":"constant","framerate":25,"futureConfigFeature":"foo","latencyMode":"realtime","alpha":"discard"} assert_equals: expected (string) "discard" but got (undefined) undefined
